pyproject.toml
[tool.poetry]
name = "flask-official-tutorial"
version = "0.1.0"
description = "Blog website built with flask for the official flask tutorial"
authors = ["David Ajawu <ajawudavid@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
email-validator = "1.1.3"
Flask = "2.0.1"
Flask-SQLAlchemy = "2.5.1"
ujson = "4.0.2"
pydantic = "^1.8.2"
python-slugify = "^5.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
